Schema update

+++ components/cc1101/fan.cpp
@@ -128,7 +128,7 @@ void CC1101Fan::control(const fan::FanCall &call) {
   }
 }
 
-void CC1101Fan::set_fan_speed(int speed) {
+void CC1101Fan::set_fan_speed(uint8_t speed) {
   ESP_LOGD("cc1101_fan", "RF called with %d while last is %d and speed assumed at %d", speed, this->LastSpeed, this->Speed);
   if (speed != this->LastSpeed ) {
     // Handle speed control
